+++ indicator.js
@@ -2,6 +2,7 @@
 // System panel indicator for Gnome Browser Switcher
 
 import GObject from 'gi://GObject';
+import Clutter from 'gi://Clutter';
 import St from 'gi://St';
 import * as PanelMenu from 'resource:///org/gnome/shell/ui/panelMenu.js';
 
@@ -11,16 +12,25 @@ import * as PanelMenu from 'resource:///org/gnome/shell/ui/panelMenu.js';
  */
 const BrowserIndicator = GObject.registerClass(
     class BrowserIndicator extends PanelMenu.Button {
-        _init(browserManager) {
+        _init(browserManager, settings) {
             super._init(0.0, 'Browser Switcher Indicator');
 
             this._browserManager = browserManager;
+            this._settings = settings;
 
             this._icon = new St.Icon({
                 icon_name: 'web-browser-symbolic',
                 style_class: 'system-status-icon',
             });
+            this._desaturationEffect = new Clutter.DesaturateEffect({ factor: 1.0 });
             this.add_child(this._icon);
+
+            this._updateDesaturation();
+            this._settings.connectObject(
+                'changed::desaturate-panel-icon',
+                () => this._updateDesaturation(),
+                this
+            );
 
             this._browserManager.watchDefaultBrowser((browserId) => {
                 this.updateIcon(browserId);
@@ -30,6 +40,21 @@ const BrowserIndicator = GObject.registerClass(
             const currentBrowser = this._browserManager.getCachedDefaultBrowser();
             if (currentBrowser)
                 this.updateIcon(currentBrowser);
+        }
+
+        /**
+         * Applies the configured visual treatment to the panel icon only.
+         * @private
+         */
+        _updateDesaturation() {
+            if (this._settings.get_boolean('desaturate-panel-icon')) {
+                this._icon.add_effect_with_name(
+                    'browser-switcher-desaturation',
+                    this._desaturationEffect
+                );
+            } else {
+                this._icon.remove_effect_by_name('browser-switcher-desaturation');
+            }
         }
 
         /**
@@ -58,7 +83,10 @@ const BrowserIndicator = GObject.registerClass(
         }
 
         destroy() {
+            this._settings.disconnectObject(this);
+            this._desaturationEffect = null;
             this._icon = null;
+            this._settings = null;
             this._browserManager = null;
             super.destroy();
         }